refactor: move  header to the component
